fastapi lastest

+++ planner/routes/events.py
@@ -1,9 +1,10 @@
-from fastapi import APIRouter, Body, HTTPException, status
+from fastapi import APIRouter, Body, HTTPException, status, Depends
 from models.events import Event, EventUpdate
 from typing import List
 
 from database.connection import Database
 from beanie import PydanticObjectId
+from auth.authenticate import authenticate
 
 event_database = Database(Event)
 event_router = APIRouter(tags=["Events"])
@@ -16,33 +17,45 @@ async def retrieve_all_events() -> List[Event]:
 @event_router.get("/{id}", response_model=Event)
 async def retrieve_event(id: PydanticObjectId) -> Event:
     event = await event_database.get(id)
-    for event in events:
-        if event.id == id:
-            return event
+    if event.id == id:
+        return event
     raise HTTPException(
         status_code=status.HTTP_404_NOT_FOUND,
         detail="Event with supplied ID does not exist"
     )
 
 @event_router.post("/new")
-async def create_event(body: Event = Body(...)) -> dict:
+async def create_event(body: Event, user: str = Depends(authenticate)) -> dict:
+    body.creator = user
     await event_database.save(body)
     return {
         "message": "Event created successfully."
     }
 
 @event_router.put("/{id}", response_model=Event)
-async def update_event(id: PydanticObjectId, body: EventUpdate) -> Event:
+async def update_event(id: PydanticObjectId, body: EventUpdate, user: str = Depends(authenticate)) -> Event:
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(
+            status_code.status.HTTP_400_BAD_REQUEST,
+            detail="Operation not allowed"
+        )
     updated_event = await event_database.update(id, body)
-    if not update_event:
+    if not updated_event:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Event with supplied ID does not exist"
         )
-    return update_event
+    return updated_event
 
 @event_router.delete("/{id}")
-async def delete_event(id: PydanticObjectId) -> dict:
+async def delete_event(id: PydanticObjectId, user: str = Depends(authenticate)) -> dict:
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Operation not allowed"
+        )
     event = await event_database.delete(id)
     if not event:
         raise HTTPException(
